feat: extract traces routes to routes/traces.py
