+++ templates/_error_panels.html
@@ -259,7 +259,7 @@
                                     <i class="bi bi-journal-text me-1"></i><span class="error-panel-btn-label">{{ logs_label }}</span>
                                 </a>
                                 {% if err.trace_id %}
-                                    <a href="{{ url_for('view_traces', trace_id=err.trace_id, from_ts=from_ts, to_ts=to_ts) }}"
+                                    <a href="{{ url_for('traces.view_traces', trace_id=err.trace_id, from_ts=from_ts, to_ts=to_ts) }}"
                                        class="btn btn-sm btn-outline-{{ trace_btn_outline }}">
                                         <i class="bi bi-diagram-3 me-1"></i><span class="error-panel-btn-label">{{ trace_label }}</span>
                                     </a>
@@ -269,7 +269,7 @@
                                    class="btn btn-sm btn-outline-{{ logs_btn_outline }}">
                                     <i class="bi bi-journal-text me-1"></i><span class="error-panel-btn-label">{{ logs_label }}</span>
                                 </a>
-                                <a href="{{ url_for('view_traces', trace_id=err.trace_id, from_ts=from_ts, to_ts=to_ts) }}"
+                                <a href="{{ url_for('traces.view_traces', trace_id=err.trace_id, from_ts=from_ts, to_ts=to_ts) }}"
                                    class="btn btn-sm btn-outline-{{ trace_btn_outline }}">
                                     <i class="bi bi-diagram-3 me-1"></i><span class="error-panel-btn-label">{{ trace_label }}</span>
                                 </a>
