+++ templates/ai.html
@@ -395,7 +395,7 @@
                   <a href="{{ url_for('logs.view_logs', trace_id=turn.trace_id, from_ts=from_ts, to_ts=to_ts) }}" class="btn btn-outline-secondary btn-sm py-0 px-2" style="font-size:.75rem;" title="View related logs">
                     <i class="bi bi-journal-text me-1"></i>Logs
                   </a>
-                  <a href="{{ url_for('view_traces', trace_id=turn.trace_id, from_ts=from_ts, to_ts=to_ts) }}" class="trace-link">
+                  <a href="{{ url_for('traces.view_traces', trace_id=turn.trace_id, from_ts=from_ts, to_ts=to_ts) }}" class="trace-link">
                     Trace {{ turn.trace_id[:12] }}…
                   </a>
                   {% endif %}
