+++ templates/base.html
@@ -740,7 +740,7 @@
         </a>
       </li>
       <li class="nav-item">
-        <a href="{{ url_for('view_traces') }}" class="nav-link {% if request.endpoint == 'view_traces' %}active{% endif %}" title="Traces">
+        <a href="{{ url_for('traces.view_traces') }}" class="nav-link {% if request.endpoint == 'traces.view_traces' %}active{% endif %}" title="Traces">
           <i class="bi bi-diagram-3"></i><span class="sb-label">Traces</span>
         </a>
       </li>
