+++ templates/custom_dashboard_view.html
@@ -580,7 +580,7 @@
   var sourceUrls = {
     logs: {{ url_for('logs.view_logs') | tojson }},
     metrics: {{ url_for('view_metrics_anomaly') | tojson }},
-    traces: {{ url_for('view_traces') | tojson }},
+    traces: {{ url_for('traces.view_traces') | tojson }},
     errors: {{ url_for('errors.view_errors') | tojson }}
   };
