+++ templates/incident_help.html
@@ -102,7 +102,7 @@
   <div class="card-body">
     <ul class="small mb-0">
       <li>From <a href="{{ url_for('errors.view_errors') }}">Errors</a> — expand an error accordion and click <strong>Incident View</strong>.</li>
-      <li>From <a href="{{ url_for('view_traces') }}">Traces</a> — open a trace detail and click <strong>Incident View</strong> in the span header.</li>
+      <li>From <a href="{{ url_for('traces.view_traces') }}">Traces</a> — open a trace detail and click <strong>Incident View</strong> in the span header.</li>
       <li>Direct URL — <code>/incident?trace_id=&lt;id&gt;</code> or <code>/incident?error_id=&lt;id&gt;</code>.</li>
     </ul>
   </div>
