+++ templates/summary.html
@@ -93,7 +93,7 @@
     </a>
   </div>
   <div class="col-6 col-md-4 col-lg-2">
-     <a href="{{ url_for('view_traces') }}" class="text-decoration-none"
+     <a href="{{ url_for('traces.view_traces') }}" class="text-decoration-none"
        data-ai-action-id="summary.nav.traces"
        data-ai-action-type="navigate"
        data-ai-handler="navigate"
