+++ templates/traces_help.html
@@ -10,7 +10,7 @@
     <h4 class="fw-semibold mb-1">Traces Help</h4>
     <p class="text-secondary mb-0">Distributed trace exploration, span drilldown, and incident correlation.</p>
   </div>
-  <a class="btn btn-outline-secondary btn-sm" href="{{ url_for('view_traces') }}" title="Back to Traces">
+  <a class="btn btn-outline-secondary btn-sm" href="{{ url_for('traces.view_traces') }}" title="Back to Traces">
     <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Traces</span>
   </a>
 </div>
